+++ supermarket-query-service/src/main/java/com/eidiko/query/config/SwaggerConfig.java
@@ -3,13 +3,12 @@ package com.eidiko.query.config;
 import io.swagger.v3.oas.annotations.OpenAPIDefinition;
 import io.swagger.v3.oas.annotations.info.Contact;
 import io.swagger.v3.oas.annotations.info.Info;
-import org.springframework.context.annotation.Configuration;
 
 @OpenAPIDefinition(
         info = @Info(
                 title = "Supermarket Query Service",
-                description = "API for supermarket query service",
-                summary = "API for supermarket query service where only get api calls are made",
+                description = "API for supermarket query service where only get api calls are made",
+                summary = "API for supermarket query service",
                 version = "v1.0",
                 contact = @Contact(
                         name = "Sanketh Shinde",
