+++ supermarket-query-service/src/main/java/com/eidiko/query/dao/EmployeeDAO.java
@@ -1,7 +1,6 @@
 package com.eidiko.query.dao;
 
 import com.eidiko.query.dto.EmployeeDTO;
-import com.eidiko.query.dto.EmployeeHierarchyDTO;
 import com.eidiko.query.exception.EmployeeNotFoundException;
 
 import java.util.List;
@@ -12,7 +11,6 @@ public interface EmployeeDAO {
 
     List<EmployeeDTO> findAll();
 
-    EmployeeHierarchyDTO findHierarchy() throws EmployeeNotFoundException;
+    List<EmployeeDTO> findByReportingTo(int id);
 
-    List<Integer> findHierarchyById(int id) throws EmployeeNotFoundException;
 }
