+++ supermarket-query-service/src/main/java/com/eidiko/query/dao/EmployeeDaoImpl.java
@@ -1,7 +1,6 @@
 package com.eidiko.query.dao;
 
 import com.eidiko.query.dto.EmployeeDTO;
-import com.eidiko.query.dto.EmployeeHierarchyDTO;
 import com.eidiko.query.exception.EmployeeNotFoundException;
 import com.eidiko.query.mapper.EmployeeMapper;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -9,10 +8,7 @@ import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 
-import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Repository
 public class EmployeeDaoImpl implements EmployeeDAO {
@@ -40,96 +36,9 @@ public class EmployeeDaoImpl implements EmployeeDAO {
     }
 
     @Override
-    public EmployeeHierarchyDTO findHierarchy() throws EmployeeNotFoundException {
-        EmployeeDTO employeeDTO = findById(1);
-        if (employeeDTO == null) {
-            throw new EmployeeNotFoundException("Employee Not Found");
-        }
-
-        // To track visited employees and avoid circular references
-        Set<Integer> visitedEmployees = new HashSet<>();
-        return buildHierarchy(employeeDTO, visitedEmployees);
-    }
-
-    @Override
-    public List<Integer> findHierarchyById(int id) throws EmployeeNotFoundException {
-        Set<Integer> visited = new HashSet<>();
-        List<Integer> hierarchyList = new ArrayList<>();
-        fetchEmployeeHierarchy(id, hierarchyList, visited);
-        return hierarchyList;
-    }
-
-    private void fetchEmployeeHierarchy(int employeeId, List<Integer> hierarchyList, Set<Integer> visited)
-            throws EmployeeNotFoundException {
-
-        if (visited.contains(employeeId)) {
-            throw new IllegalStateException(
-                    "Circular dependency detected in employee hierarchy for Employee ID: " + employeeId
-            );
-        }
-
-        visited.add(employeeId);
-
-        EmployeeDTO employeeDTO = findById(employeeId);
-        EmployeeDTO manager = findById(employeeDTO.getReportingTo());
-
-        hierarchyList.add(manager.getId());
-        if (manager.getDesignation().contains("Sales Supervisor")) {
-            return;
-        }
-
-        if (manager.getReportingTo() != 0) {
-            fetchEmployeeHierarchy(manager.getId(), hierarchyList, visited);
-        }
-    }
-
-    private EmployeeHierarchyDTO buildHierarchy(EmployeeDTO employeeDTO, Set<Integer> visitedEmployees) {
-        // Check if the employee has already been visited to avoid infinite recursion
-        if (visitedEmployees.contains(employeeDTO.getId())) {
-            // Return null if this employee has already been processed
-            return null;
-        }
-
-        // Mark the current employee as visited
-        visitedEmployees.add(employeeDTO.getId());
-
-        return createEmployeeHierarchy(employeeDTO, visitedEmployees);
-    }
-
-    private EmployeeHierarchyDTO createEmployeeHierarchy(EmployeeDTO employeeDTO, Set<Integer> visitedEmployees) {
-        // Create EmployeeHierarchyDTO for the employee
-        EmployeeHierarchyDTO employeeHierarchyDTO = new EmployeeHierarchyDTO();
-        employeeHierarchyDTO.setId(employeeDTO.getId());
-        employeeHierarchyDTO.setName(employeeDTO.getName());
-        employeeHierarchyDTO.setEmail(employeeDTO.getEmail());
-        employeeHierarchyDTO.setDesignation(employeeDTO.getDesignation());
-        employeeHierarchyDTO.setRole(employeeDTO.getRole());
-        employeeHierarchyDTO.setPhoneNumber(employeeDTO.getPhoneNumber());
-        employeeHierarchyDTO.setJoiningDate(employeeDTO.getJoiningDate());
-        employeeHierarchyDTO.setSalary(employeeDTO.getSalary());
-
-        // Get the subordinates (employees who report to this one)
-        List<EmployeeDTO> subordinates = findSubordinates(employeeDTO.getId());
-
-        // Recursively build hierarchy for each subordinate
-        List<EmployeeHierarchyDTO> subordinateHierarchyList = new ArrayList<>();
-        for (EmployeeDTO subordinate : subordinates) {
-            EmployeeHierarchyDTO subordinateHierarchy = buildHierarchy(subordinate, visitedEmployees);
-            if (subordinateHierarchy != null) {
-                subordinateHierarchyList.add(subordinateHierarchy);
-            }
-        }
-
-        // Set the subordinates in the employee hierarchy DTO
-        employeeHierarchyDTO.setEmployees(subordinateHierarchyList);
-        return employeeHierarchyDTO;
-    }
-
-    // Helper method to find subordinates (employees reporting to a specific employee)
-    private List<EmployeeDTO> findSubordinates(int managerId) {
-        // Assuming this is where you'd fetch subordinates from the database based on the manager's ID.
+    public List<EmployeeDTO> findByReportingTo(int id) {
         String query = "SELECT * FROM employees WHERE reporting_to = ?";
-        return postgresqlJdbcTemplate.query(query, new EmployeeMapper(), managerId);
+        return postgresqlJdbcTemplate.query(query, new EmployeeMapper(), id);
     }
 
 }
